contact refactoring and features adjustemnt
